Extract TypicalLoanBook and add Bike support (#120)
* Extract TypicalLoanBook to its own file

* Update location of TypicalLoanBook for everyone

* Add typical bikes to TypicalLoanBook

* Add typical bikes to typicalLoansLoanBook.xml

* Rename typicalLoansLoanBook.xml to typicalLoanBook.xml

* Correct checkstyles
