+++ src/test/java/seedu/address/model/LoanBookTest.java
@@ -6,8 +6,8 @@ import static org.junit.Assert.assertTrue;
 import static seedu.address.logic.commands.CommandTestUtil.VALID_ADDRESS_BOB;
 import static seedu.address.logic.commands.CommandTestUtil.VALID_TAG_HUSBAND;
 import static seedu.address.testutil.TypicalBikes.BIKE1;
+import static seedu.address.testutil.TypicalLoanBook.getTypicalLoanBook;
 import static seedu.address.testutil.TypicalLoans.ALICE;
-import static seedu.address.testutil.TypicalLoans.getTypicalLoanBook;
 
 import java.util.Arrays;
 import java.util.Collection;
