+++ src/test/java/seedu/address/storage/XmlLoanBookStorageTest.java
@@ -2,10 +2,10 @@ package seedu.address.storage;
 
 import static org.junit.Assert.assertEquals;
 import static org.junit.Assert.assertFalse;
+import static seedu.address.testutil.TypicalLoanBook.getTypicalLoanBook;
 import static seedu.address.testutil.TypicalLoans.ALICE;
 import static seedu.address.testutil.TypicalLoans.HOON;
 import static seedu.address.testutil.TypicalLoans.IDA;
-import static seedu.address.testutil.TypicalLoans.getTypicalLoanBook;
 
 import java.io.IOException;
 import java.nio.file.Path;
